+++ test/integration-test.js
@@ -2099,6 +2099,10 @@ title']{display:block}",
       'a{background:url(1.png) 0 0/28px 28px}'
     ]
   }),
+  'background size with -properties.backgroundSizeMerging': cssContext({
+    'standard': 'div{background:url(image.png) no-repeat center;background-size:cover}',
+    'prefix': 'div{-webkit-background:url(image.png) no-repeat center;-webkit-background-size:cover}'
+  }, { compatibility: '-properties.backgroundSizeMerging' }),
   'multiple backgrounds': cssContext({
     'should not produce longer values': 'p{background:no-repeat;background-position:100% 0,0 100%,100% 100%,50% 50%}'
   }),